Move PupilDataBisector to player_methods
