+++ pupil_src/shared_modules/video_export/plugins/world_video_exporter.py
@@ -123,7 +123,6 @@ def _export_world_video(
     from vis_light_points import Vis_Light_Points
     from vis_polyline import Vis_Polyline
     from vis_watermark import Vis_Watermark
-    from pupil_producers import PupilDataBisector
 
     PID = str(os.getpid())
     logger = logging.getLogger(__name__ + " with pid: " + PID)
@@ -233,7 +232,7 @@ def _export_world_video(
                 for serialized in initializers["data"]
             ]
 
-        g_pool.pupil_positions = PupilDataBisector.from_init_dict(
+        g_pool.pupil_positions = pm.PupilDataBisector.from_init_dict(
             pre_computed_eye_data["pupil"]
         )
         g_pool.gaze_positions = pm.Bisector(**pre_computed_eye_data["gaze"])
